refactor: move upload to ShellSessionsController for all shell types
Upload is now tied to the shell token instead of the project, so it
works for project shells, workbench shells, and add-on shells. The
upload endpoint uses the token's cluster/namespace/pod directly,
removing the need for pod_name validation against running pods.

Bracket paste (auto-pasting path into terminal) is only enabled for
workbench sessions. Upload hint shown on all terminals.
