+++ app/controllers/projects/workbenches_controller.rb
@@ -1,9 +1,6 @@
 class Projects::WorkbenchesController < Projects::BaseController
   before_action :require_development_environment
 
-  MAX_UPLOAD_SIZE = 10.megabytes
-  UPLOAD_DIR = "/tmp/uploads"
-
   def show
     all_pods = fetch_pods
     @pods = all_pods.select { |pod| pod.status.phase == "Running" }
@@ -24,50 +21,10 @@ class Projects::WorkbenchesController < Projects::BaseController
     end
   end
 
-  def upload
-    file = params[:file]
-
-    if file.blank?
-      return render json: { error: "No file provided" }, status: :unprocessable_entity
-    end
-
-    if file.size > MAX_UPLOAD_SIZE
-      return render json: { error: "File too large (max #{MAX_UPLOAD_SIZE / 1.megabyte}MB)" }, status: :unprocessable_entity
-    end
-
-    pod_name = params[:pod_name]
-    namespace = @project.namespace
-    running_pods = fetch_pods.select { |pod| pod.status.phase == "Running" }
-    valid_pod = running_pods.any? { |pod| pod.metadata.name == pod_name }
-
-    unless valid_pod
-      return render json: { error: "Invalid pod" }, status: :unprocessable_entity
-    end
-
-    filename = sanitize_filename(file.original_filename)
-    remote_path = "#{UPLOAD_DIR}/#{filename}"
-
-    kubectl = K8::Kubectl.new(active_connection)
-    container_flag = [ "-c", @project.name ]
-
-    kubectl.call([ "exec", "-n", namespace, pod_name ] + container_flag + [ "--", "mkdir", "-p", UPLOAD_DIR ])
-    kubectl.call([ "cp", file.tempfile.path, "#{namespace}/#{pod_name}:#{remote_path}" ] + container_flag)
-
-    render json: { path: remote_path }
-  rescue Cli::CommandFailedError => e
-    render json: { error: "Upload failed: #{e.message}" }, status: :unprocessable_entity
-  end
-
   private
 
   def require_development_environment
     redirect_to project_path(@project), alert: "Workbench is only available for development environments." unless @project.development_environment?
-  end
-
-  def sanitize_filename(name)
-    # Strip path components, replace unsafe chars, prepend timestamp to avoid collisions
-    base = File.basename(name).gsub(/[^\w.\-]/, "_")
-    "#{Time.current.to_i}_#{base}"
   end
 
   def fetch_pods
