+++ app/views/projects/processes/shell.html.erb
@@ -5,7 +5,6 @@
     namespace: @namespace,
     back_path: project_processes_path(@project),
     project_name: @project.name,
-    cluster: @project.cluster,
-    upload_url: upload_project_workbench_path(@project)
+    cluster: @project.cluster
   %>
 <% end %>
