+++ app/views/projects/workbenches/show.html.erb
@@ -9,7 +9,6 @@
       back_path: project_path(@project.child_development_environment.parent_project),
       project_name: @project.name,
       cluster: @project.cluster,
-      upload_url: upload_project_workbench_path(@project),
       bracket_paste: true,
       show_sessions: true %>
   <% else %>
